Update codebase with faster dBdi and cleaned MAIN
This update cleans up the codebase a lot and also speeds up one of the
slower symbolic functions. Specifically, some changes include:
- Reduced number of functions in active_functions to about half the
number of files. Only functions that were not used were removed.
- Added folder containing all experimental data used in this work.
- Added a MAIN.m file that guides the user through the workflow. It
still does not call the Switching Time Optimization scripts, however.
- Reduced the number of MAIN scripts to simplify workflow and eliminate
experimental scripts.
- Unified VAR files into one VARIABLES.m file.
- Sped up dBdi function to be numeric, not symbolic. It creates minor
variation in results as compared to symbolic output, but that is
primarily visible near the coils, where the swimmer does not operate. I
think it is unlikely that this will have a negative effect on code's
operation.

I still need to go through the Switching Time Optimization codes and
clean them up. I will do this soon as they will be used in the next
stage of development on this project. For now they run, but I have
noticed that I only successfully optimize the switching times if the
starting guess gets the swimmer to the correct position on the vertical
axis on the first try. The algorithm appears to only be able to optimize
the swimmer's position in the horizontal, not vertical direction.